[YS-257] refactor: 모집 대상 필터링 상수 및 타입 분리
