+++ src/app/home/home.types.ts
@@ -1,3 +1,5 @@
+import { GENDER } from './home.constants';
+
 export type Area =
   | '전국'
   | '서울'
@@ -50,3 +52,5 @@ export interface SeoulRegion {
   name: SeoulArea;
   count: number;
 }
+
+export type GenderValue = (typeof GENDER)[number]['value'];
